fix compilation under linux

+++ AQUtil/aq/FileMapper.cpp
@@ -19,7 +19,7 @@ FileMapper::~FileMapper()
 int FileMapper::read(void * buffer, size_t offset, size_t len) 
 {
   assert(this->m_fd);
-  assert(offset < std::numeric_limits<long>::max());
+  assert(offset < std::numeric_limits<size_t>::max());
   fseek(this->m_fd, static_cast<long>(offset), SEEK_SET);
   fread(buffer, 1, len, this->m_fd);
   if (feof(this->m_fd))
